+++ kurtosis-module/impl/services/indexer.ts
@@ -2,13 +2,11 @@ import { FilesArtifactUUID, EnclaveContext, ServiceID, ContainerConfig, Containe
 import * as log from "loglevel";
 import { Result, ok, err } from "neverthrow";
 import { EXEC_COMMAND_SUCCESS_EXIT_CODE } from "../consts";
-import { ContainerConfigSupplier } from "../near_module";
 import { getPrivateAndPublicUrlsForPortId, ServiceUrl } from "../service_url";
-import * as path from "path";
 import { waitForPortAvailability } from "../service_port_availability_checker"
 
 const SERVICE_ID: ServiceID = "indexer-node"
-const IMAGE: string = "kurtosistech/near-indexer-for-explorer:c07fb7b";
+const IMAGE: string = "kurtosistech/near-indexer-for-explorer:2d66461";
 const RPC_PRIVATE_PORT_NUM: number = 3030;
 const RPC_PUBLIC_PORT_NUM: number = 8332;
 const RPC_PORT_ID = "rpc";